Alert management started   -> internship status update alert   -> showing alerts on top bar
 version heroku

+++ src/main/java/com/storactive/stg/repository/AlertRepo.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AlertRepo extends JpaRepository<Alert, Integer>, JpaSpecificationExecutor<Alert> {
-    long countByStatusEquals(int status);
+    long countByStatusEquals(short status);
 }
